proxy/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
requests>=2.28.0
python-multipart>=0.0.6